test-requirements.txt
coverage==3.7.1
nose-cov
testfixtures
nose==1.3.7
mock==3.0.5
flake8

# For integration tests
pytest==4.6.3
pyyaml>=4.2b1
cloudify-utilities-plugins-sdk>=0.0.77
